# [FIX] - Address memory issue where entire source code is retained in memory preventing GC