--- a/packages/core/feature-flags/src/index.ts
+++ b/packages/core/feature-flags/src/index.ts
@@ -200,6 +200,12 @@ export const DEFAULT_FEATURE_FLAGS = {
    */
   preserveUnstableSingleFileOutputInCss:
     process.env.ATLASPACK_BUILD_ENV === 'test',
+
+  /**
+   * When enabled, avoid retaining sourcesContent in memory during transformation.
+   * Source contents will be inlined only during final map stringify if needed.
+   */
+  omitSourcesContentInMemory: false,
 };
 
 export type FeatureFlags = typeof DEFAULT_FEATURE_FLAGS;