--- a/packages/core/integration-tests/test/sourcemaps.ts
+++ b/packages/core/integration-tests/test/sourcemaps.ts
@@ -1659,24 +1659,24 @@ describe('sourcemaps', function () {
   }
 
   it('does not retain sourcesContent in transformers when omitSourcesContentInMemory is enabled', async function () {
-    // Enable flag via env; integration test harness reads feature flags from options.featureFlags
-    const prevFlag =
-      process.env.ATLASPACK_FEATURE_FLAG_omitSourcesContentInMemory;
-    process.env.ATLASPACK_FEATURE_FLAG_omitSourcesContentInMemory = 'true';
-    try {
-      await bundle(
-        path.join(__dirname, '/integration/sourcemap-existing/index.js'),
-      );
-      const outputs = await outputFS.readdir(distDir);
-      assert(outputs.some((f) => f.endsWith('.map')));
-      // Purely a sanity check that the map is generated with this flag enabled.
-      const mapPath = outputs.find((f) => f.endsWith('.map'));
-      const map = JSON.parse(
-        await outputFS.readFile(path.join(distDir, String(mapPath))),
-      );
-      assert(map.mappings && map.mappings.length > 0);
-    } finally {
-      process.env.ATLASPACK_FEATURE_FLAG_omitSourcesContentInMemory = prevFlag;
-    }
+    await bundle(
+      path.join(__dirname, '/integration/sourcemap-existing/index.js'),
+      {featureFlags: {omitSourcesContentInMemory: true}},
+    );
+    const outputs = await outputFS.readdir(distDir);
+    const mapPath = outputs.find((f) => f.endsWith('.map'));
+    const map = JSON.parse(
+      await outputFS.readFile(path.join(distDir, String(mapPath))),
+    );
+
+    assert(outputs.some((f) => f.endsWith('.map')));
+    assert(map.mappings && map.mappings.length > 0);
+    assert(Array.isArray(map.sourcesContent));
+    assert.strictEqual(map.sourcesContent.length, map.sources.length);
+    assert(
+      map.sourcesContent.some(
+        (s: unknown) => typeof s === 'string' && s.length > 0,
+      ),
+    );
   });
 });