--- a/packages/transformers/js/src/JSTransformer.ts
+++ b/packages/transformers/js/src/JSTransformer.ts
@@ -585,7 +585,9 @@ export default new Transformer({
                         // @ts-expect-error TS2345
                         map.extends(originalMap);
                       } else {
-                        map.setSourceContent(asset.filePath, code.toString());
+                        if (!getFeatureFlag('omitSourcesContentInMemory')) {
+                          map.setSourceContent(asset.filePath, code.toString());
+                        }
                       }
                     }