## Motivation

Currently builds on very large projects require machines with a lot of RAM; This is because G.C isn't able to clear memory effectively because a reference to the underlying source code is being maintained longer than it needs to be; Leading to the entire codebase staying in memory.

## Changes
- Summary: Introduce omitSourcesContentInMemory feature flag to stop retaining full source text in sourcesContent throughout transforms. This reduces in-process memory without changing default behavior or sourcemap outputs.
- Default behavior: Unchanged (flag is OFF). Source maps and build outputs are identical.
- When enabled: Avoids pre-populating/storing sourcesContent during transform; mappings and source paths remain correct; packager still inlines or references sources per config.

Memory Analysis (three-js benchmark, V3 mode, 1 run)

baseline:
```
~44.2s Mark-Compact 20.4 (50.2) -> 12.5 (53.5) MB
~45.3s Mark-Compact 21.8 (36.0) -> 12.4 (54.0) MB
```

with flag:
```
~17.7s Mark-Compact 11.5 (28.0) -> 9.6 (27.5) MB
~19.7s Mark-Compact 11.9 (44.5) -> 9.3 (44.5) MB
```


## Checklist

- [ x] Existing or new tests cover this change
- [ x] There is a changeset for this change, or one is not required