@marcins 2025-09-21
What's the story with sources content and the native asset graph? (i.e. "Atlaspack v3") Does it work the same way there? Could / should the same optimisation be applied?