@shanshrew 2025-09-22
> What's the story with sources content and the native asset graph? (i.e. "Atlaspack v3") Does it work the same way there? Could / should the same optimisation be applied?

PackageRUnner.generateSourceMap is used by both v2 and v3; The optimisation should work across all modes.