@yamadapc 2025-09-22
There are extra changes that can be made to the native build to get the gain in CI