@yamadapc 2025-09-22
Actually it might already be omitted there:

* https://github.com/swc-project/swc/blob/72a53a0de5ddf5d5163369b18d7c90c0baeb4280/crates/swc_common/src/source_map.rs#L1510-L1516
* https://github.com/atlassian-labs/atlaspack/blob/ed0a5b2913a54f168913aa7361bfc7f3c1d50ffa/packages/transformers/js/core/src/lib.rs#L737-L746